projects/retail-simulator/seed/customer/requirements.txt
Faker
psycopg2-binary==2.9.1
